% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_algorithms.R
\name{check_algorithms}
\alias{check_algorithms}
\title{check_algorithms}
\usage{
check_algorithms(algorithms)
}
\arguments{
\item{algorithms}{list containing a list for each algorithm. Each sublist
contains 1) name  and 2) function}
}
\value{

}
\description{
check_algorithms
}
