% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_metric.R
\name{check_metric}
\alias{check_metric}
\title{check and process the supplied metric and metric name}
\usage{
check_metric(metric, metric.name)
}
\arguments{
\item{metric}{either "cor" or a function evaluating 
the similarity of two vectors}

\item{metric.name}{character containing name of the metric}
}
\value{
list with entries metric, metric.name
}
\description{
check and process the supplied metric and metric name
}
