% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_fun.R
\name{cost_fun}
\alias{cost_fun}
\title{cost function to minimize for fitting curves to correlations for different
datasets}
\usage{
cost_fun(p, temp_df, ct)
}
\arguments{
\item{p}{named vector of parameters that are optimized}

\item{temp_df}{data frame containing columns 'celltype', 'error', 
'correlation'}

\item{ct}{character, cell type for which curve should be fit}
}
\value{
residual (loss to be minimized)
}
\description{
cost function to minimize for fitting curves to correlations for different
datasets
}
