% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_boxplots.R
\name{create_boxplots}
\alias{create_boxplots}
\title{create boxplots of deconvolution results for each algorithm}
\usage{
create_boxplots(results.df, celltype.order = NULL, algorithm.order = NULL)
}
\arguments{
\item{results.df}{data frame as returned by prepare_data}

\item{celltype.order}{character vector of cell types
specifying the plotting order}

\item{algorithm.order}{character vector of algorithm names
specifying the plotting order (left to right)}
}
\value{
list containing list of ggplot objects (plots),
ordered by celltype.order (if supplied)
}
\description{
create boxplots of deconvolution results for each algorithm
}
