% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bulks.R
\name{create_bulks}
\alias{create_bulks}
\title{simulate bulk data by summing over single-cell data}
\usage{
create_bulks(
  exprs,
  pheno,
  n.bulks = 500,
  include.in.bulks = NULL,
  fraction.per.bulk = 0.1,
  sum.to.count = TRUE
)
}
\arguments{
\item{exprs}{non-negative numeric, scRNA-seq profiles as columns, 
features as rows}

\item{pheno}{data.frame, phenotype data containing
cell type labels for the expression matrix,
label column must be named 'cell_type',
ordering of cells must be the same as in exprs}

\item{n.bulks}{integer, the number of bulks to be created, defaults to 500}

\item{include.in.bulks}{vector of strings, cell types to be used for bulk 
simulation; if not supplied, all will be used}

\item{fraction.per.bulk}{positive numeric < 1, fraction of samples to be randomly
drawn for each bulk; default 0.1}

\item{sum.to.count}{boolean, should all bulks be normalized
to a fixed total count number? default TRUE}
}
\value{
list with 
   - "bulks": matrix containing bulk expression profiles (features x bulks)
   - "props" matrix containing quantities (cell type x bulks)
   - "sub.props": TODO
}
\description{
simulate bulk data by summing over single-cell data
}
