% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_mat.R
\name{create_cov_mat}
\alias{create_cov_mat}
\title{create a matrix of covariances between cell type quantities in C}
\usage{
create_cov_mat(C)
}
\arguments{
\item{C}{non-negative numeric matrix (cell type x bulk) of cellular 
composition}
}
\value{
matrix of covariances (cell type x cell type)
}
\description{
create a matrix of covariances between cell type quantities in C
}
