% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_score_plots.R
\name{create_fit_plots}
\alias{create_fit_plots}
\title{This function creates fit plots and plots of offsets and errors (scores)
from the output of `score_algorithms`}
\usage{
create_fit_plots(score.results)
}
\arguments{
\item{score.results}{list of data frames as read from h5 result file; 
each data frame is one output of `score_algorithms`}
}
\value{
list containing score plots:\cr
1) fits: list containing one fit plot for each repetition\cr
2) offset_plot: boxplot of fitted offset for each algorithm and celltype\cr
3) error_plot: boxplot of fitted error for each algorithm and celltype\cr
4) combined_plot: barplot of (offset + error) for each algorithm and celltype
}
\description{
This function creates fit plots and plots of offsets and errors (scores)
from the output of `score_algorithms`
}
