% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lineplots.R
\name{create_lineplots}
\alias{create_lineplots}
\title{create lineplots of deconvolution results for different gene sets}
\usage{
create_lineplots(
  results.df,
  metric = "cor",
  genesets = NULL,
  available.features = NULL,
  celltype.order = NULL,
  algorithm.order = NULL
)
}
\arguments{
\item{results.df}{data frame as returned by prepare_data}

\item{metric}{character string specifying the evaluation metric; default 'cor'}

\item{genesets}{list of gene sets (character vectors)}

\item{available.features}{character vector containing names of available features}

\item{celltype.order}{character vector of cell types specifying the plotting order}

\item{algorithm.order}{character vector of algorithm names specifying the plotting order}
}
\value{
list containing two elements:  

1) runtime.plot  

2) list containing lineplot of scores for each celltypes
}
\description{
create lineplots of deconvolution results for different gene sets
}
