% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scatterplots.R
\name{create_scatterplots}
\alias{create_scatterplots}
\title{create scatterplots of deconvolution results for all algorithms}
\usage{
create_scatterplots(
  results.list,
  real.props = NULL,
  training.pheno = NULL,
  real = FALSE,
  celltype.order = NULL,
  algorithm.order = NULL
)
}
\arguments{
\item{results.list}{list of results as returned by deconvolute}

\item{real.props}{matrix containing real proportions for each cell type
'(cell type x bulk); optional, needed only if results.list does not contain
'bulk.props'}

\item{training.pheno}{pheno data for the single-cell training set; optional}

\item{real}{boolean; are the deconvoluted bulks real or artificial}

\item{celltype.order}{character vector of cell types
specifying the plotting order}

\item{algorithm.order}{character vector of algorithm names
specifying the plotting order}
}
\value{
list containing one plot per algorithm, ordered by algorithm.order
}
\description{
create scatterplots of deconvolution results for all algorithms
}
