% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_score_tables.R
\name{create_score_tables}
\alias{create_score_tables}
\title{extract matrices of averaged offsets and errors from
score simulation results}
\usage{
create_score_tables(score.results)
}
\arguments{
\item{score.results}{list of data frames as read from h5 result file; 
each data frame is one output of `score_algorithms`}
}
\value{
list with two entries:\cr
1) offsets: matrix of offsets (algorithm x cell type)\cr
2) errors: matrix of errors (algorithm x cell type)
}
\description{
extract matrices of averaged offsets and errors from
score simulation results
}
