% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sig_matrix.R
\name{create_sig_matrix}
\alias{create_sig_matrix}
\title{create a signature matrix for usage with CIBERSORT or DeconRNASeq
according to description in Newman et. al}
\usage{
create_sig_matrix(
  exprs,
  pheno,
  exclude.celltypes = NULL,
  max.genes = NULL,
  optimize = TRUE,
  split.data = TRUE
)
}
\arguments{
\item{exprs}{non-negative numeric matrix,}

\item{pheno}{data.frame, phenodata corresponding to
expression data, same ordering of samples/cells; cell type
label column must be named 'cell_type'}

\item{exclude.celltypes}{character vector, defaults to
c("malignant", "not_annotated", "unassigned"). Cells
with these labels are not included in the signature matrix}

\item{max.genes}{integer, maximum number of genes to be selected
for each cell type; the actual number of genes in the
signature matrix will be larger. If no value is given,
max.genes = NULL and all genes will be selected}

\item{optimize}{boolean, default TRUE. If TRUE, the
optimal number of genes to be selected for each cell type will be
determined between 1 and 'max.genes' by optimizing the
condition number of the signature matrix.
If FALSE, for each cell type the top 'max.genes' genes will be taken.}

\item{split.data}{logical, should the training data be split for
reference profile creation and optimization? default: TRUE}
}
\value{
numeric matrix, holding reference profiles in its column, 
features in its rows
}
\description{
create a signature matrix for usage with CIBERSORT or DeconRNASeq
according to description in Newman et. al
}
