% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluate_deconvolution}
\alias{evaluate_deconvolution}
\title{evaluate deconvolution performance using different metrics}
\usage{
evaluate_deconvolution(real, estimate)
}
\arguments{
\item{real}{matrix containing real quantities of cell types (cell type x proportion)}

\item{estimate}{matrix containing estimations of cell type quantities (cell type x proportion)}
}
\value{
list containing a score for each metric: correlation (cor), mad (mean absolute difference)
and rmsd (root mean squared distance)
}
\description{
evaluate deconvolution performance using different metrics
}
