% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fine_coarse_subtype_benchmark.R
\name{fine_coarse_subtype_benchmark}
\alias{fine_coarse_subtype_benchmark}
\title{fine and coarse subtype benchmark function}
\usage{
fine_coarse_subtype_benchmark(
  sc.counts = cll.exprs,
  sc.pheno = cll.pheno,
  subtype.pattern = "subtype",
  cell.type.column = "cell_type",
  sample.name.column = "sample.name",
  n.cluster.sizes = c(1, 2, 4, 8),
  verbose = TRUE,
  algorithm.list = list(list(algorithm = run_dtd, name = "DTD"), list(algorithm =
    run_least_squares, name = "Least_Squares")),
  patient.column = NULL,
  n.bulks = 500
)
}
\arguments{
\item{sc.counts}{}

\item{sc.pheno}{}

\item{subtype.pattern}{}

\item{cell.type.column}{}

\item{sample.name.column}{}

\item{verbose}{}

\item{algorithm.list}{}
}
\value{

}
\description{
# basically, there are 3 different things to compare. 
1. correlation per algorithm on the deep C 
2. correlation per algorithm on the accumulated cursory C 
3. special scenario where subtype 1 is in X, and subtype 2 is in Y
}
\details{
1. and 2. can be done in one go, 3. needs more prepration.
}
