% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_data.R
\name{fit_data}
\alias{fit_data}
\title{fit the correlation model of DAB to deconvolution results}
\usage{
fit_data(results, datasets, nbulks = NULL, nsets = NULL)
}
\arguments{
\item{results}{List containing one sublist for bulk dataset 
(each simulated with a different variance in cellular composition). 
Each sublist contains deconvolution results (matrix, cell type x bulk) for
all algorithms.}

\item{nbulks}{integer > 0, number of bulks in each of the `nsets` data sets}

\item{nsets}{integer > 0, number of datasets to be drawn for deconvolution}

\item{dataset}{The bulk datasets that were deconvoluted. List containing 
bulk counts and proportions as created by `create_bulks` for different 
variances}
}
\value{
data frame containing results 
(correlations, model parameters, predictions ...)
}
\description{
fit the correlation model of DAB to deconvolution results
}
