% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset_benchmark.R
\name{geneset_benchmark}
\alias{geneset_benchmark}
\title{perform deconvolution of simulated bulks on different gene sets}
\usage{
geneset_benchmark(
  training.exprs,
  training.pheno,
  test.exprs,
  test.pheno,
  genesets,
  algorithms,
  bulk.data,
  n.repeats = 3,
  exclude.from.signature = NULL,
  verbose = FALSE,
  cell.type.column = "cell_type",
  patient.column = "patient"
)
}
\arguments{
\item{training.exprs}{matrix containing single-cell expression profiles
(training set, one cell per column)}

\item{training.pheno}{data frame containing phenotype data of the
single-cell training set. Has to contain column `cell.type.column`}

\item{test.exprs}{matrix containing single-cell expression profiles
(test set, one cell per column)}

\item{test.pheno}{data frame containing phenotype data of the
single-cell test set. Has to contain column `cell.type.column`}

\item{genesets}{list of gene sets (character vectors)}

\item{algorithms}{List containing a list for each algorithm.
Each sublist contains 1) name, 2) function and 3) model}

\item{bulk.data}{list with two entries:\cr
1) bulks - matrix containing expression data of the bulks
(one bulk per column)\cr
2) props - matrix containing the true fractions of cell types within the
bulks (cell type x bulk)}

\item{n.repeats}{integer determining the number of times deconvolution
should be repeated for each algorithm, default 3}

\item{exclude.from.signature}{character vector containing cell types to be
excluded from the signature matrix. If not specified, all will be used.}

\item{verbose}{logical, default FALSE}

\item{cell.type.column}{string, which column of 'training.pheno'/'test.pheno'
holds the cell type information? default "cell_type"}

\item{patient.column}{string, which column of 'pheno'
holds the patient information; optional, default "patient"}
}
\value{
list containing deconvolution results for all algorithms
for all genesets
}
\description{
perform deconvolution of simulated bulks on different gene sets
}
