% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5_read_mat.R
\name{h5_read_mat}
\alias{h5_read_mat}
\title{read a simple matrix from hdf5 file}
\usage{
h5_read_mat(filename)
}
\arguments{
\item{filename}{character specifying the file to read from}
}
\value{
matrix read from file
}
\description{
read a simple matrix from hdf5 file
}
