% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5_write_mat.R
\name{h5_write_mat}
\alias{h5_write_mat}
\title{write a simple matrix to a hdf5 file}
\usage{
h5_write_mat(m, filename, filepath = "")
}
\arguments{
\item{m}{matrix with > 0 rows and > 0 named columns}

\item{filename}{filename of hdf5 file to be written to}

\item{filepath}{optional, path where the file should be stored.
Only necessary, if the directory does not yet exist, otherwise it can
be supplied as part of filename}
}
\value{
NULL, write to file
}
\description{
write a simple matrix to a hdf5 file
}
