% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_genes.R
\name{marker_genes}
\alias{marker_genes}
\title{select marker genes following description in Baron et al., 2016}
\usage{
marker_genes(exprs, pheno, sig.types = NULL, cell.type.column = "cell_type")
}
\arguments{
\item{exprs}{matrix containing single cell profiles as columns}

\item{pheno}{phenotype data corresponding to the expression matrix.
Has to contain single cell labels in a column named `cell.type.column`}

\item{sig.types}{character vector containing cell types for which
marker genes should be selected. Default is all of them.}

\item{cell.type.column}{string, which column of 'pheno'
holds the cell type information?}
}
\value{
list containing a vector of marker genes for each cell type
}
\description{
select marker genes following description in Baron et al., 2016
}
