% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_check_model.R
\name{pipeline_check_model}
\alias{pipeline_check_model}
\title{pipeline check model function}
\usage{
pipeline_check_model(DTD.model)
}
\arguments{
\item{DTD.model}{: any R object, 
this object gets evaluated, if it is/can be a valid DTD.model}
}
\value{
boolean, wheter 'DTD.model' is valid
}
\description{
Given a DTD model, this function checks wheter the model is valid
}
