% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cond_num.R
\name{plot_cond_num}
\alias{plot_cond_num}
\title{create condition number plots for deconvolution results}
\usage{
plot_cond_num(results.df, algorithm.order = NULL)
}
\arguments{
\item{results.df}{data frame containing results as returned by prepare_data}

\item{algorithm.order}{character vector of algorithm names
specifying the plotting order}
}
\value{
list containing 3 plots:\cr
1) cond_num_plot - barplot of average condition numbers\cr
2) cond_vs_score - scatter plot of
average condition number vs. score for each algorithm\cr
3) variation_plot - scatter plot of standard deviation of
score vs. standard deviation of condition number
}
\description{
create condition number plots for deconvolution results
}
