% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_present_algos.R
\name{present_algos}
\alias{present_algos}
\title{read results written by previous benchmarks and return a list
of algorithms already present in the data}
\usage{
present_algos(target_dir, name_pattern)
}
\arguments{
\item{target_dir}{character, directory from which results are to be read}

\item{name_pattern}{character, pattern by which relevant files are recognized}
}
\value{
character vector containing algorithm names
}
\description{
read results written by previous benchmarks and return a list
of algorithms already present in the data
}
