% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bseqsc_wrapper.R
\name{run_bseqsc}
\alias{run_bseqsc}
\title{deconvolute given bulks with BSEQ-sc using single cell data}
\usage{
run_bseqsc(
  exprs,
  pheno,
  bulks,
  exclude.from.signature = NULL,
  max.genes = NULL,
  optimize = TRUE,
  split.data = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{exprs}{non negative numeric matrix containing single cell profiles
as columns and features as rows}

\item{pheno}{data.frame, with 'nrow(pheno)' must equal 'ncol(exprs)'. 
Has to contain single cell labels in a column named 'cell_type'}

\item{bulks}{matrix containing bulk expression profiles as columns}

\item{exclude.from.signature}{vector of strings of cell types not to be
included in the signature matrix}

\item{max.genes}{numeric, maximum number of genes that will be included in 
the signature for each celltype}

\item{optimize}{boolean, should the signature matrix be optimized by
condition number? If FALSE, max.genes genes will be used}

\item{split.data}{boolean, should the training data be split for signature
matrix creation? If TRUE, 10% of the data will be used to build
the signature matrix and the rest will be used to estimate the optimal
features}

\item{verbose}{boolean}
}
\value{
list with four entries: 
1) est.props - matrix containing for each bulk the
estimated fractions of the cell types contained
2) sig.matrix - effective signature matrix used by the algorithm (features x cell types)
}
\description{
deconvolute given bulks with BSEQ-sc using single cell data
}
