% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_cibersort_lm22_jerbyarnon.R
\name{run_cibersort_lm22_jerbyarnon}
\alias{run_cibersort_lm22_jerbyarnon}
\title{deconvolute given bulks with CIBERSORT using LM22 model}
\usage{
run_cibersort_lm22_jerbyarnon(
  exprs,
  pheno,
  bulks,
  exclude.from.signature = NULL,
  max.genes = 500,
  cell.type.column = "cell_type",
  patient.column = NULL,
  scale.cpm = FALSE,
  model = NULL,
  model_exclude = NULL
)
}
\arguments{
\item{exprs}{non negative numeric matrix containing single cell profiles
as columns and features as rows}

\item{pheno}{data.frame, with 'nrow(pheno)' must equal 'ncol(exprs)'.
Has to contain single cell labels in a column named 'cell_type'}

\item{bulks}{matrix containing bulk expression profiles as columns}

\item{exclude.from.signature}{vector of strings of cell types not to be
included in the signature matrix}

\item{max.genes}{numeric, maximum number of genes that will be included in
the signature for each celltype, default 500}

\item{cell.type.column}{string, which column of 'pheno'
holds the cell type information? default "cell_type"}

\item{patient.column}{string, which column of 'pheno'
holds the patient information; optional, default NULL}

\item{scale.cpm}{boolean, unused}

\item{verbose}{boolean}
}
\value{
list with four entries:
1) est.props - matrix containing for each bulk the
estimated fractions of the cell types contained
2) sig.matrix - effective signature matrix used by the algorithm
(features x cell types)
3) model - NULL\cr
}
\description{
deconvolute given bulks with CIBERSORT using LM22 model
}
