% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_cibersort_random.R
\name{run_cibersort_random_genes}
\alias{run_cibersort_random_genes}
\title{deconvolute given bulks with CIBERSORT using single cell data
(random gene selection)}
\usage{
run_cibersort_random_genes(
  exprs,
  pheno,
  bulks,
  exclude.from.signature = NULL,
  max.genes = 500,
  cell.type.column = "cell_type",
  patient.column = NULL,
  scale.cpm = FALSE,
  model = NULL,
  model_exclude = NULL
)
}
\arguments{
\item{exprs}{non negative numeric matrix containing single cell profiles
as columns and features as rows}

\item{pheno}{data.frame, with 'nrow(pheno)' must equal 'ncol(exprs)'.
Has to contain single cell labels in a column named 'cell_type'}

\item{bulks}{matrix containing bulk expression profiles as columns}

\item{exclude.from.signature}{vector of strings of cell types not to be
included in the signature matrix}

\item{max.genes}{numeric, maximum number of genes that will be included in
the signature for each celltype, default 500}

\item{cell.type.column}{string, which column of 'pheno'
holds the cell type information? default "cell_type"}

\item{patient.column}{string, which column of 'pheno'
holds the patient information; optional, default NULL}

\item{scale.cpm}{boolean, scale single-cell profiles to CPM? default FALSE}

\item{model}{model for CIBERSORT deconvolution as returned by this wrapper,
default NULL}

\item{model_exclude}{character vector, cell type(s) to exclude
from the supplied pre-trained model, default NULL}

\item{verbose}{boolean}
}
\value{
list with four entries:
1) est.props - matrix containing for each bulk the
estimated fractions of the cell types contained\cr
2) sig.matrix - effective signature matrix used by the algorithm
(features x cell types)\cr
3) model - list containing reference.X (signature matrix)\cr
}
\description{
deconvolute given bulks with CIBERSORT using single cell data
(random gene selection)
}
