% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_music.R
\name{run_music}
\alias{run_music}
\title{deconvolute given bulks with MuSiC using single cell data}
\usage{
run_music(
  exprs,
  pheno,
  bulks,
  exclude.from.signature = NULL,
  max.genes = NULL,
  cell.type.column = "cell_type",
  patient.column = "patient",
  scale.cpm = FALSE,
  model = NULL,
  model_exclude = NULL
)
}
\arguments{
\item{exprs}{non negative numeric matrix containing single cell profiles
as columns and features as rows}

\item{pheno}{data.frame, with 'nrow(pheno)' must equal 'ncol(exprs)'.
Has to contain single cell labels in a column named 'cell_type'}

\item{bulks}{matrix containing bulk expression profiles as columns}

\item{exclude.from.signature}{vector of strings of cell types not to be
included in the signature matrix}

\item{max.genes}{unused, MuSiC uses all genes}

\item{cell.type.column}{string, which column of 'pheno'
holds the cell type information? default "cell_type"}

\item{patient.column}{string, which column of 'pheno'
holds the patient information; optional, default "patient"}

\item{scale.cpm}{boolean, scale single-cell profiles to CPM? default FALSE}

\item{model}{pre-trained model for MuSiC deconvolution
as returned by this wrapper, default NULL}

\item{model_exclude}{character vector, cell type(s) to exclude
from the supplied pre-trained model, default NULL}

\item{verbose}{boolean, default FALSE}
}
\value{
list with three entries:
1) est.props - matrix containing for each bulk the
estimated fractions of the cell types contained\cr
2) sig.matrix - effective signature matrix used by the algorithm
(features x cell types)\cr
3) model - list containing learned parameters for music.basic function call
}
\description{
deconvolute given bulks with MuSiC using single cell data
}
