% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_counts.R
\name{scale_to_count}
\alias{scale_to_count}
\title{rescale expression profiles to sum to a fixed count number}
\usage{
scale_to_count(exprs, count = NULL)
}
\arguments{
\item{exprs}{expression matrix containing profiles in columns}

\item{count}{numeric, the number the counts in each profile should sum up to:
default NULL, count = nrow(exprs)}
}
\value{
rescaled expression matrix
}
\description{
rescale expression profiles to sum to a fixed count number
}
