% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_algorithms.R
\name{score_algorithms}
\alias{score_algorithms}
\title{fit correlation models to deconvolution results obtained on simulated bulks}
\usage{
score_algorithms(
  counts,
  pheno,
  bulk_counts,
  bulk_pheno,
  bulk_cell_types = NULL,
  exclude_from_signature = NULL,
  column_names = list(cell.type.column = "celltype", patient.column = "patient",
    sample.column = "sample"),
  algorithm_list,
  nrep = 5,
  nsets = 4,
  nbulks = 250,
  datasets = NULL
)
}
\arguments{
\item{counts}{non-negative numeric matrix of single-cell counts
(genes x cells)}

\item{pheno}{data frame containing phenotype information corresponding
to the cells in `counts`}

\item{bulk_counts}{non-negative numeric matrix of single-cell counts
(genes x cells); profiles to be used for bulk creation}

\item{bulk_pheno}{data frame containing phenotype information corresponding
to the cells in `bulk_counts`}

\item{bulk_cell_types}{character vector containing names of cell types to
be included in simulated bulks; default NULL}

\item{exclude_from_signature}{character vector containing names of cell
types to be excluded from deconvolution models; default NULL}

\item{column_names}{list with 3 entries: \cr
1) cell.type.column: string, name of column in pheno data frames containing
cell types\cr
2) patient.column: string, name of column in pheno data frames containing
patient names / sample of origin of the single-cell profiles\cr
3) sample.column: string, name of the column containing names of the
profiles contained in the count matrices (must correspond to
colnames(counts / bulk_counts))}

\item{algorithm_list}{input list of algorithms; see `benchmark` for info}

\item{nrep}{integer > 0, number of repetitions (for averaging and
error estimation)}

\item{nsets}{integer > 0, number of datasets to be drawn for deconvolution}

\item{nbulks}{integer > 0, number of bulks in each of the `nsets` data sets}

\item{datasets}{list of simulated bulk datasets as returned by this function;
optional, default NULL}
}
\value{
list containing two elements:\cr
1) results - data frame containing fit results
2) datasets - list containing bulk datasets for all variances
}
\description{
fit correlation models to deconvolution results obtained on simulated bulks
}
