% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtype_plot.R
\name{subtype_plot}
\alias{subtype_plot}
\title{create lineplot of algorithm performance across different cell type
granularities}
\usage{
subtype_plot(results.df)
}
\arguments{
\item{results.df}{data frame as returned by prepare_data}
}
\value{
list containing\cr
1) lineplot of all algorithms\cr
2) lineplots of individual algorithms\cr
3) lineplot of coarse correlations after fine deconvolution
}
\description{
create lineplot of algorithm performance across different cell type
granularities
}
