% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undmc.R
\name{undmc}
\alias{undmc}
\title{Return the hex code and RGB values of a given DMC embroidery floss}
\usage{
undmc(dmc, visualize = TRUE)
}
\arguments{
\item{dmc}{DMC floss identifier. Usually the floss number (e.g. 210), or the name if it does not commonly have a number (e.g. "Ecru").}

\item{visualize}{Whether to visualize the DMC floss color. Defaults to TRUE.}
}
\description{
Return the hex code and RGB values of a given DMC embroidery floss
}
\examples{
undmc("Ecru")
undmc(310)
undmc(c(210, 211))
}
