% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5_write_data.R
\name{write_data}
\alias{write_data}
\title{write scRNA-Seq profiles and pheno data, RNA-Seq bulk data and cell type 
proportions within the bulks in format used by \link{benchmark} to hdf5 file}
\usage{
write_data(
  sc.counts = NULL,
  sc.pheno = NULL,
  bulk.counts = NULL,
  bulk.props = NULL,
  sub.props = NULL,
  filename
)
}
\arguments{
\item{sc.counts}{numeric matrix, with feature as rows and scRNA-Seq
profiles as columns}

\item{sc.pheno}{data frame, with scRNA-Seq profiles as rows, and pheno
entries as rows}

\item{bulk.counts}{numeric matrix, with features as rows bulk measurements
as columns}

\item{bulk.props}{numeric matrix containing cell type proportions, cell types as rows, 
bulks as columns}

\item{sub.props}{numeric matrix containing cell type proportions of simulated subtypes, 
cell types as rows, bulks as columns}

\item{filename}{string, where should the data be stored?}
}
\value{
NULL, function saves into ‘filename‘
}
\description{
write scRNA-Seq profiles and pheno data, RNA-Seq bulk data and cell type 
proportions within the bulks in format used by \link{benchmark} to hdf5 file
}
