% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5_write_misc.R
\name{write_misc_input}
\alias{write_misc_input}
\title{write some input data of the benchmark to hdf5}
\usage{
write_misc_input(genesets, algorithm.names, function.call, grouping, filename)
}
\arguments{
\item{genesets}{list of string vectors}

\item{algorithm.names}{cahracter vector; names of the deconvolution algorithms used}

\item{function.call}{call object}

\item{grouping}{numeric vector grouping samples in test and training set}

\item{filename}{string, where should the data be stored?}
}
\value{
NULL, function saves into ‘filename‘
}
\description{
write some input data of the benchmark to hdf5
}
