% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5_write_results.R
\name{write_result_list}
\alias{write_result_list}
\title{write result list as returned by deconvolute and
simulation functions recursively}
\usage{
write_result_list(results.all, filename, group = NULL)
}
\arguments{
\item{results.all}{list of lists, as generated by the deconvolution and
simulation functions}

\item{filename}{string, where should the data be stored?}

\item{group}{string, group name for the hdf5 file (NULL for top level)}
}
\value{
NULL, function saves into ‘filename‘
}
\description{
write result list as returned by deconvolute and
simulation functions recursively
}
