% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_subtype_pheno.R
\name{add_subtype_pheno}
\alias{add_subtype_pheno}
\title{add subtype pheno}
\usage{
add_subtype_pheno(
  sc.counts,
  sc.pheno,
  cell.type.column = "cell_type",
  sample.name.column = "sample.name",
  new.subtype.column = "subtype",
  hclust.obj = NA,
  n.clusters = c(1, 2, 4, 8),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{sc.counts}{count matrix, features as rows, scRNA-Seq profiles as columns}

\item{sc.pheno}{data.frame. scRNA-Seq profiles as rows. Must have 
'cell.type.column' and 'sample.name.column'}

\item{cell.type.column}{string, column of 'sc.pheno' holding the 
input cell type labels. Within these entries, the clustering is done.}

\item{sample.name.column}{string, column of the 'colnames(sc.counts)'}

\item{new.subtype.column}{string, pattern of the new column, where the subtype
information is stored}

\item{hclust.obj}{'hclust' object, or NA. If you already calculated a hclust, 
object pass it here. Alternatively, we call the get_hclust function.}

\item{verbose}{logical, should information about the process be printed.}

\item{...}{arguments that are passed to get_hclust}

\item{avg.profiles.per.subcluster.vec}{numeric vector, indicates how many new 
subtype pheno entries are generated 
(=> 'length(avg.profiles.per.subcluster.vec)') and - in average - how many
profiles are in a subtype.}
}
\value{

}
\description{
Use a hierarchical clustering of the input labels, 
to generate finer subtype labels. 
The basic steps are: 
   1. use the get_hclust function to hierarchically cluster the data
    within the cell types. (or use the provided hclust.obj)
   2. cluster the data multiple times, with different depth 
   (as indicated by 'avg.profiles.per.subcluster.vec')
   3. provide the clustering labels as columns in sc.pheno, and return it.
}
