% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_subtypes.R
\name{assign_subtypes}
\alias{assign_subtypes}
\title{simulate subtypes of given cell types using t-SNE and k-means clustering}
\usage{
assign_subtypes(sc.counts, sc.pheno, sub.list, celltypecol = "cell_type", ...)
}
\arguments{
\item{sc.counts}{non-negative numeric matrix with features as rows, and 
scRNA-Seq profiles as columns. 'ncol(sc.counts)' must equal 'nrow(sc.pheno)'}

\item{sc.pheno}{data frame with scRNA-Seq profiles as rows, and pheno entries
in columns. 'nrow(sc.pheno)' must equal 'ncol(sc.counts)'}

\item{sub.list}{named list containing the number of subtypes to be simulated for each cell type}

\item{celltypecol}{string, column name in 'sc.pheno' that holds the cell type
per scRNA-Seq profile}

\item{...}{additional parameters that get passed to Rtsne()}
}
\value{
list containing two entries:  

1) sc.pheno - input pheno data with added column 'subtype'  

2) tsne.embed - t-SNE embedding (output of Rtsne) that was used to generate
the subtypes
}
\description{
simulate subtypes of given cell types using t-SNE and k-means clustering
}
