% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark}
\alias{benchmark}
\title{main function of the benchmark}
\usage{
benchmark(
  sc.counts,
  sc.pheno,
  real.counts,
  real.props,
  benchmark.name,
  grouping,
  input.algorithms = NULL,
  simulation.bulks = FALSE,
  simulation.genes = FALSE,
  simulation.samples = FALSE,
  simulation.subtypes = FALSE,
  genesets = NULL,
  metric = "cor",
  repeats = 3,
  temp.dir = NULL,
  exclude.from.bulks = NULL,
  exclude.from.signature = NULL,
  n.bulks = 1000,
  cpm = TRUE,
  verbose = FALSE,
  n.subtypes = 3
)
}
\arguments{
\item{sc.counts}{non-negative numeric matrix with features as rows, and 
scRNA-Seq profiles as columns. 'ncol(sc.counts)' must equal 'nrow(sc.pheno)'}

\item{sc.pheno}{data frame with scRNA-Seq profiles as rows, and pheno entries
in columns. 'nrow(sc.pheno)' must equal 'ncol(sc.counts)'}

\item{real.counts}{non-negative numeric matrix, with features as rows, and 
bulk RNA-Seq profiles as columns. 'ncol(sc.counts)' must equal 
'nrow(real.props)'}

\item{real.props}{non-negative numeric matrix, with cell types as rows, 
and bulk RNA-Seq profiles.}

\item{benchmark.name}{string}

\item{grouping}{factor with 2 levels, and 'length(grouping)' must be 
'ncol(sc.counts)'. Assigns each scRNA-Seq profile to either 
test or train cohort.}

\item{input.algorithms}{list containing a list for each algorithm. 
Each sublist contains 1) name  and 2) function
For predefined algorithms it is sufficient to supply only the name instead of the sublist
#TODO: examples!!!}

\item{simulation.bulks}{boolean, should deconvolution of simulated bulks be
performed, default FALSE}

\item{simulation.genes}{boolean, should deconvolution of simulated bulks with 
be predefined genesets be performed, default FALSE}

\item{simulation.samples}{boolean, should deconvolution of simulated bulks with
varying number of random training profiles be performed, default FALSE}

\item{simulation.subtypes}{boolean, should deconvolution of simulated bulks with
artificial subtypes of given cell types be performed, default FALSE}

\item{genesets}{list of string vector, must match 'rownames(sc.counts)'}

\item{metric}{string, must match one of 'c("cor", "mad", "rmsd")' 
#TODO: keep this up to date}

\item{repeats}{numeric > 0}

\item{temp.dir}{string, directory where data, and benchmarks get stored.}

\item{exclude.from.bulks}{vector of strings, cell types that should not be 
predicted by the algorithms}

\item{exclude.from.signature}{}

\item{n.bulks}{numeric > 0}

\item{cpm}{boolean, should the sc profiles and bulks be scaled to counts
per million}

\item{verbose}{boolean, should progress information be printed to the screen}

\item{n.subtypes}{numeric > 0}

\item{missing.algorithm}{character specifying the algorithm for which a model
with incomplete reference should be tested; default DTD}

\item{missing.celltype}{character specifiying cell type; compare DTD model with
reference matrix not containing this cell type to complete models}
}
\value{
NULL, results are stored via hdf5 to 'temp.dir'
}
\description{
main function of the benchmark
}
\examples{
see either 'working_example.R', or 'working_example_fast.R'
}
