% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_algorithms.R
\name{check_algorithms}
\alias{check_algorithms}
\title{check a list of algorithms for compatitbility with \link{benchmark}}
\usage{
check_algorithms(algorithms)
}
\arguments{
\item{algorithms}{list containing a list for each algorithm. Each sublist
contains \cr 1) name: character  \cr 2) algorithm: function}
}
\description{
check a list of algorithms for compatitbility with \link{benchmark}
}
