% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bulks.R
\name{create_bulks}
\alias{create_bulks}
\title{simulate bulk data by summing over single-cell data}
\usage{
create_bulks(
  exprs,
  pheno,
  cell.type.column = "cell_type",
  n.bulks = 500,
  include.in.bulks = NULL,
  n.profiles.per.bulk = 1000,
  sum.to.count = TRUE,
  frac = 5,
  switch.prob = 0
)
}
\arguments{
\item{exprs}{non-negative numeric, scRNA-seq profiles as columns, 
features as rows}

\item{pheno}{data.frame, phenotype data containing
cell type labels for the expression matrix,
must contain 'cell.type.column',
ordering of cells must be the same as in exprs}

\item{cell.type.column}{string, which column of 'pheno'
holds the cell type information?}

\item{n.bulks}{integer, the number of bulks to be created, defaults to 500}

\item{include.in.bulks}{vector of strings, cell types to be used for bulk 
simulation; if not supplied, all will be used}

\item{n.profiles.per.bulk}{positive numeric, number of samples to be randomly
drawn for each bulk; default 1000}

\item{sum.to.count}{boolean, should all bulks be normalized
to a fixed total count number? default TRUE}

\item{frac}{numeric >= 1; determines the variance of cell type proportions within the simulated bulks;
higher d means more extreme distributions}

\item{switch.prob}{fraction of bulks sampled with inverse cell type quantity}
}
\value{
list with 
   - "bulks": matrix containing bulk expression profiles (features x bulks)
   - "props" matrix containing quantities (cell type x bulks)
}
\description{
simulate bulk data by summing over single-cell data
}
