% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lineplots.R
\name{create_lineplots}
\alias{create_lineplots}
\title{create lineplots of deconvolution results for different gene sets}
\usage{
create_lineplots(
  results.df,
  metric = "cor",
  metric.name = NULL,
  genesets = NULL,
  available.features = NULL,
  celltype.order = NULL,
  algorithm.order = NULL
)
}
\arguments{
\item{results.df}{data frame as returned by prepare_data}

\item{metric}{evaluation metric; either string 'cor' (default) or a function}

\item{metric.name}{string, name of the evaluation metric used; not needed if metric is a string ("cor"). If metric is a function and metric.name
is NULL, the default will be "custom metric"}

\item{genesets}{list of gene sets (character vectors)}

\item{available.features}{character vector containing names of available features}

\item{celltype.order}{character vector of cell types specifying the plotting order}

\item{algorithm.order}{character vector of algorithm names specifying the plotting order}
}
\value{
list containing two elements:  

1) runtime.plot  

2) list containing lineplot of scores for each celltypes
}
\description{
create lineplots of deconvolution results for different gene sets
}
