% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sig_matrix.R
\name{create_sig_matrix}
\alias{create_sig_matrix}
\title{create a signature matrix for usage with CIBERSORT or DeconRNASeq
according to description in Newman et. al}
\usage{
create_sig_matrix(
  exprs,
  pheno,
  exclude.celltypes = NULL,
  max.genes = NULL,
  cell.type.column = "cell_type"
)
}
\arguments{
\item{exprs}{non-negative numeric matrix,}

\item{pheno}{data.frame, phenodata corresponding to
expression data, same ordering of samples/cells; cell type
label column must be named `cell.type.column`}

\item{exclude.celltypes}{character vector, defaults to
c("malignant", "not_annotated", "unassigned"). Cells
with these labels are not included in the signature matrix}

\item{max.genes}{integer, maximum number of genes to be selected
for each cell type; the actual number of genes in the
signature matrix will be larger. If no value is given,
max.genes = NULL and all genes will be selected as maximum}

\item{cell.type.column}{string, which column of 'pheno'
holds the cell type information? default "cell_type"}
}
\value{
numeric matrix, holding reference profiles in its column,
features in its rows
}
\description{
create a signature matrix for usage with CIBERSORT or DeconRNASeq
according to description in Newman et. al
}
