% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmc.R
\name{dmc}
\alias{dmc}
\title{Find the closest DMC embroidery floss for a given colour}
\usage{
dmc(colour, n = 1, visualize = TRUE, method = c("euclidean",
  "cie1976", "cie94", "cie2000", "cmc"))
}
\arguments{
\item{colour}{A hex code to identify the colour, e.g. "#EE8726"}

\item{n}{Number of DMC floss colours to return. Defaults to 1.}

\item{visualize}{Whether to visualize \code{colour} versus the closest DMC floss colour(s). Defaults to TRUE.}

\item{method}{Method for comparing colours. Defaults to "euclidean".}
}
\description{
Find the closest DMC embroidery floss for a given colour
}
\examples{
dmc("#EE8726")
dmc("#EE8726", n = 5)
dmc("#EE8726", n = 5, method = "cie94")
dmc("#EE8726", visualize = FALSE)
}
