% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scores.R
\name{evaluation_plot}
\alias{evaluation_plot}
\title{create score plot}
\usage{
evaluation_plot(
  results.df,
  title = NULL,
  real.props = NULL,
  celltype.order = NULL,
  algorithm.order = NULL
)
}
\arguments{
\item{results.df}{data frame as returned by prepare_data}

\item{title}{character, plot title}

\item{real.props}{non-negative numeric matrix, with cell types as rows,
and bulk RNA-Seq profiles.}

\item{celltype.order}{character vector of cell types specifying
the plotting order}

\item{algorithm.order}{character vector of algorithm names
specifying the plotting order (left to right)}
}
\description{
create score plot
}
