% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fine_coarse_subtype_benchmark.R
\name{fine_coarse_subtype_benchmark}
\alias{fine_coarse_subtype_benchmark}
\title{fine and coarse subtype benchmark function}
\usage{
fine_coarse_subtype_benchmark(
  sc.counts,
  sc.pheno,
  algorithm.list,
  subtype.pattern = "subtype",
  cell.type.column = "cell_type",
  sample.name.column = "sample.name",
  n.clusters = c(1, 2, 4, 8),
  verbose = FALSE,
  patient.column = "patient",
  n.bulks = 500,
  repeats = 3
)
}
\arguments{
\item{sc.counts}{count matrix, features as rows,
scRNA-Seq profiles as columns}

\item{sc.pheno}{data.frame. scRNA-Seq profiles as rows. Must have
'cell.type.column' and 'sample.name.column'}

\item{algorithm.list}{List containing a list for each algorithm.
Each sublist contains 1) name, 2) function and 3) model}

\item{subtype.pattern}{character,
string by which subtype column is recognized; default "subtype"}

\item{cell.type.column}{string, column of 'sc.pheno' holding the
input cell type labels. Within these entries, the clustering is done.}

\item{sample.name.column}{string, column of the 'colnames(sc.counts)'}

\item{n.clusters}{integer vector of clustering depths (number of subclusters
created for each cell type), default c(1, 2, 4, 8).
This means that in the finest clustering, each celltype will be split in
8 subtypes, in the next step each will be split in 4 subtypes, ...}

\item{verbose}{logical, should information about the process be printed?}

\item{patient.column}{string, which column of 'pheno'
holds the patient information; optional, default "patient"}

\item{n.bulks}{numeric > 0, number of bulks to simulate. default 500}

\item{repeats}{numeric > 0,
number of repetitions for each algorithm. default: 3}
}
\value{
list containing deconvolution results for
different cell type granularities
}
\description{
# basically, there are 3 different things to compare.
1. correlation per algorithm on the deep C
2. correlation per algorithm on the accumulated cursory C
3. special scenario where subtype 1 is in X, and subtype 2 is in Y
}
\details{
1. and 2. can be done in one go, 3. needs more prepration.
}
