% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset_benchmark.R
\name{geneset_benchmark}
\alias{geneset_benchmark}
\title{perform deconvolution of simulated bulks on different gene sets}
\usage{
geneset_benchmark(
  training.exprs,
  training.pheno,
  test.exprs,
  test.pheno,
  genesets,
  algorithms,
  bulk.data,
  n.repeats,
  exclude.from.signature = NULL,
  verbose = FALSE,
  split.data = FALSE
)
}
\arguments{
\item{training.pheno}{data frame containing phenotype data of the single-cell training set. Has to contain column "cell_type"}

\item{test.pheno}{data frame containing phenotype data of the single-cell test set. Has to contain column "cell_type"}

\item{genesets}{list of gene sets (character vectors)}

\item{algorithms}{List containing a list for each algorithm. Each sublist contains 1) name  and 2) function}

\item{bulk.data}{list with two entries:
1) bulks - matrix containing expression data of the bulks (one bulk per column)
2) props - matrix containing the true fractions of cell types within the bulks (cell type x bulk)}

\item{n.repeats}{integer determining the number of times deconvolution should be repeated for each algorithm}

\item{exclude.from.signature}{character vector containing cell types to be excluded from the signature matrix.
If not specified, all will be used.}

\item{verbose}{logical, default FALSE}

\item{split.data}{logical, if TRUE (default) then 10% of the training data will be used for reference profile creation and
the rest for feature selection/optimization}

\item{training.expr}{matrix containing single-cell expression profiles (training set, one cell per column)}

\item{test.expr}{matrix containing single-cell expression profiles (test set, one cell per column)}
}
\value{
list containing deconvolution results for all algorithms for all genesets
}
\description{
perform deconvolution of simulated bulks on different gene sets
}
