% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all.R
\name{plot_all}
\alias{plot_all}
\title{load results from temp_dir, create all possible plots and save them into
temp_dir/report_plots}
\usage{
plot_all(temp_dir, genesets, features)
}
\arguments{
\item{temp_dir}{string, directory containing benchmark results}

\item{genesets}{list of cahracter vectors specifying gene sets used}

\item{features}{character vector containing all available features in data}
}
\value{
list containing two elements:\cr
1) celltype_order: vector containing cell types in same order as displayed in plot on real data\cr
2) celltype_order_simulated: same as above, for simulated data
}
\description{
load results from temp_dir, create all possible plots and save them into
temp_dir/report_plots
}
