% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{prepare_data}
\usage{
prepare_data(results.all, metric = "cor")
}
\arguments{
\item{results.all}{list of lists, as generated by the 'benchmark()' function}

\item{metric}{string, must match one of 'c("cor", "mad", "rmsd")'}
}
\value{
data frame containing information about every deconvolution result in the dataset
}
\description{
prepare_data
}
