% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{convert result lists from benchmark() into data frame for plotting}
\usage{
prepare_data(results.all)
}
\arguments{
\item{results.all}{list of lists, as generated by the 'benchmark()' function}
}
\value{
data frame containing information about every deconvolution result
in the dataset
}
\description{
convert result lists from benchmark() into data frame for plotting
}
