% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5_read_result.R
\name{read_result_list}
\alias{read_result_list}
\title{read_result_list}
\usage{
read_result_list(filename, content = NULL, groupname = NULL)
}
\arguments{
\item{filename}{string, which file should be read in?}

\item{content}{h5 file content; used instead of filename in case of recursive function call}

\item{groupname}{name of the current group}
}
\value{
list
}
\description{
read_result_list
}
