% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_results.R
\name{render_results}
\alias{render_results}
\title{render markdown report}
\usage{
render_results(temp.dir, metric = "cor")
}
\arguments{
\item{temp.dir}{directory containing the benchmark data}

\item{metric}{character specifying the metric, default 'cor'}
}
\description{
render markdown report
}
