% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_results.R
\name{render_results}
\alias{render_results}
\title{render markdown report for data produced by \link{benchmark}}
\usage{
render_results(temp.dir, benchmark.name = "", celltype.col = "cell_type")
}
\arguments{
\item{temp.dir}{directory containing the benchmark data. \cr
Usually \code{paste(temp.dir,"/",benchmark.name,sep="")},
where \code{temp.dir} and \code{benchmark.name}
are parameters of \link{benchmark}.}

\item{benchmark.name}{string, name of the benchmark to be included in
report filename}

\item{celltype.col}{string, name of the column in sc.pheno that contains
cell type information; default "cell_type"}
}
\description{
render markdown report for data produced by \link{benchmark}
}
