% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undmc.R
\name{undmc}
\alias{undmc}
\title{Return the hex code and RGB values of a given DMC embroidery floss}
\usage{
undmc(dmc)
}
\arguments{
\item{dmc}{DMC floss identifier (usually the number)}
}
\description{
Return the hex code and RGB values of a given DMC embroidery floss
}
\examples{
undmc("Ecru")
undmc(310)
}
